% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color}
\alias{color}
\title{Creates an object of class 'color'}
\usage{
color(..., maxVal = 255)
}
\arguments{
\item{...}{A character containing a color name, a hexcode, a c(r,g,b) vector or
three single numbers representing r,g,b respectively}

\item{maxVal}{Value assigned to represent maximum amount of light. Defaults to 255}
}
\description{
Creates an object of class 'color'
}
