% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_blindness_red.R
\docType{data}
\name{color_blindness_red}
\alias{color_blindness_red}
\title{Arabidopsis QTL data on gravitropism}
\format{An object of class \code{matrix} with 9 rows and 3 columns.}
\source{
\href{http://web.archive.org/web/20081014161121/http://www.colorjack.com/labs/colormatrix/}
}
\usage{
data(color_blindness_red)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\keyword{datasets}
