% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_preview.R
\name{color_preview}
\alias{color_preview}
\title{Previsualizes colors in a plot window}
\usage{
color_preview(...)
}
\arguments{
\item{...}{Colors to be displayed.}
}
\description{
Previsualizes colors in a plot window
}
\examples{
color_preview(colorblind("red", type="protanopia"), colorblind("green", type="protanopia"))
color_preview("#000000", "#333333", "#666666", "#999999", "#bbbbbb", "#eeeeee")
}
