% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complementary.R
\name{complementary}
\alias{complementary}
\title{Returns the complementary color to that it receives as input}
\usage{
complementary(...)
}
\arguments{
\item{...}{Color whose complementary is to be calculated. Hexcode, c(r,g,b), color 
name as character and "color" class object are all supported.}
}
\description{
Returns the complementary color to that it receives as input
}
