% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\title{Calculates the minimum contrast between col and the colors specified in ...}
\usage{
contrast(col, ...)
}
\arguments{
\item{col}{The color we want to compare. Hexcode, c(r,g,b), color 
name as character and "color" class object are all supported.}

\item{...}{All other colors that col will be compared to. Hexcode, c(r,g,b), color 
name as character and "color" class object are all supported.}
}
\description{
Calculates the minimum contrast between col and the colors specified in ...
}
