% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_color2.R
\name{df_color2}
\alias{df_color2}
\title{Appends a column named "color" to a data frame containing hex codes (ready
for usage in a plot) The values are determined by mixing two gradients that depend on two columns
specified by the user.Therefore, the output is able to represent the value of two columns together.
With "rgb" method, simultaneous low values for both columns return darker colors.
With "hsv" method and "value" as second_parameter, darker points represent low values for the second variable. The first variable is represented by a blue-green-red gradient
With "hsv" method and "saturation" as second_parameter, attenuated colors represent low values for the second variable. The first variable is represented by a blue-green-red gradient}
\usage{
df_color2(df, column1, column2, output_name = "color",
  max_color1 = "red", max_color2 = "green", method = "rgb",
  second_param = "saturation", min_saturation = 0.2, min_value = 0.4)
}
\arguments{
\item{df}{Data frame containing the original data.}

\item{column1}{A string containing the name of the column that will determine color's first gradient.}

\item{column2}{A string containing the name of the column that will determine color's second gradient.}

\item{output_name}{Name for the column in which the output will be stored. Defaults to "color"}

\item{max_color1}{In case "rgb" method is used, color to be assigned to the row with the highest value in column1. Defaults to red}

\item{max_color2}{In case "rgb" method is used, color to be assigned to the row with the highest value in column2. Defaults to green.}

\item{method}{Type of gradient. Allows both "rgb" and "hsv". Defaults to "rgb"}

\item{second_param}{In case "hsv" method is used, choose between "saturation" and "value" to be used as second parameter.}

\item{min_saturation}{In case "hsv" method is used, sets minimum allowed value for "saturation". Data points with too low saturation may not appear visible. Defaults to 0.2}

\item{min_value}{In case "hsv" method is used, sets minimum allowed value for "value". Data points with too low values may appear black regardless of the "hue" parameter. Defaults to 0.4}
}
\value{
A data frame identical to df, except for an extra column (whose name is given by output_name) containing hexcodes.
}
\description{
Appends a column named "color" to a data frame containing hex codes (ready
for usage in a plot) The values are determined by mixing two gradients that depend on two columns
specified by the user.Therefore, the output is able to represent the value of two columns together.
With "rgb" method, simultaneous low values for both columns return darker colors.
With "hsv" method and "value" as second_parameter, darker points represent low values for the second variable. The first variable is represented by a blue-green-red gradient
With "hsv" method and "saturation" as second_parameter, attenuated colors represent low values for the second variable. The first variable is represented by a blue-green-red gradient
}
\examples{
plot(mtcars$hp, mtcars$disp, col=df_color2(mtcars, "drat", "wt",method="rgb")$color, pch=18, cex=1.5)
plot(mtcars$hp, mtcars$disp, col=df_color2(mtcars, "drat", "wt",method="rgb", max_color1="green", max_color2="blue")$color, pch=18, cex=1.5)
plot(mtcars$hp, mtcars$disp, col=df_color2(mtcars, "drat", "wt", second_param = "saturation", method="hsv")$color, pch=18, cex=1.5)
plot(mtcars$hp, mtcars$disp, col=df_color2(mtcars, "drat", "wt", second_param ="value", method="hsv")$color, pch=18, cex=1.5)
}
