% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_color3.R
\name{df_color3}
\alias{df_color3}
\title{Appends a column named "color" to a data frame containing hex codes (ready
for usage in a plot) Three different columns from a data frame provide amounts for red, green
and blue light, which combine into a color for each observation, thus allowing the user to 
represent simultaneaously up to three columns using color.}
\usage{
df_color3(df, column1, column2, column3, output_name = "color",
  min_value1 = min(df[, column1]), max_value1 = max(df[, column1]),
  min_value2 = min(df[, column2]), max_value2 = max(df[, column2]),
  min_value3 = min(df[, column3]), max_value3 = max(df[, column3]),
  avoid_white = TRUE)
}
\arguments{
\item{df}{Data frame containing the original data.}

\item{column1}{A string containing the name of the column that will determine the output's amount of red light}

\item{column2}{A string containing the name of the column that will determine the output's amount of green light.}

\item{column3}{A string containing the name of the column that will determine the output's amounto of blue light.}

\item{output_name}{Name for the column in which the output will be stored. Defaults to "color"}

\item{min_value1}{Allows the user to customize the value for column1 that gets no red light assigned. Defaults to the minium of all observations.}

\item{max_value1}{Allows the user to customize the value for column1 that gets maximum red light assigned. Defaults to the maximum of all observations.}

\item{min_value2}{Allows the user to customize the value for column1 that gets no green light assigned. Defaults to the minium of all observations.}

\item{max_value2}{Allows the user to customize the value for column1 that gets maximum green light assigned. Defaults to the maximum of all observations.}

\item{avoid_white}{When set to TRUE, corrects colors that would be hard to distinguish in a white background.}

\item{min_value2}{Allows the user to customize the value for column1 that gets no blue light assigned. Defaults to the minium of all observations.}

\item{max_value2}{Allows the user to customize the value for column1 that gets maximum blue light assigned. Defaults to the maximum of all observations.}
}
\description{
Appends a column named "color" to a data frame containing hex codes (ready
for usage in a plot) Three different columns from a data frame provide amounts for red, green
and blue light, which combine into a color for each observation, thus allowing the user to 
represent simultaneaously up to three columns using color.
}
