% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots_as_list.R
\name{dots_as_list}
\alias{dots_as_list}
\title{Standardizes the format of ... arguments. If a list is given, no
modification is made. Otherwise, a list is created with all arguments
in ...}
\usage{
dots_as_list(...)
}
\arguments{
\item{...}{Input to be converted into a list.}
}
\description{
Standardizes the format of ... arguments. If a list is given, no
modification is made. Otherwise, a list is created with all arguments
in ...
}
