% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_into_interval.R
\name{fit_into_interval}
\alias{fit_into_interval}
\title{Corrects a number to be in a given interval}
\usage{
fit_into_interval(x, min = 0, max = 255)
}
\arguments{
\item{x}{Number to be corrected}

\item{min}{Lower bound of the interval. If x is lower than min, the function will return min.}

\item{max}{Upper bound of the interval. If x is higher than max, the function will return max.}
}
\description{
Corrects a number to be in a given interval
}
