% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paint.R
\name{get_paint}
\alias{get_paint}
\title{Gets YMCK specifications (paint composition) from a "color" object}
\usage{
get_paint(col, maxVal = 255)
}
\arguments{
\item{col}{Color whose complementary is to be calculated. Hexcode, c(r,g,b), color
name as character and "color" class object are all supported.}
}
\value{
A vector containing col's YMCK (paint) specifications.
}
\description{
Gets YMCK specifications (paint composition) from a "color" object
}
\examples{
get_paint("green")
get_paint(c(255,0,255))
}
