% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rgb.R
\name{get_rgb}
\alias{get_rgb}
\title{Gets RGB specifications from a "color" object}
\usage{
get_rgb(col)
}
\arguments{
\item{col}{An object of class color.}
}
\description{
Gets RGB specifications from a "color" object
}
