% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex}
\alias{hex}
\title{Gets hexcode from a "color" object.}
\usage{
hex(col)
}
\arguments{
\item{col}{Color whose complementary is to be calculated. Hexcode, c(r,g,b), color
name as character and "color" class object are all supported.}
}
\value{
A string containing col's hex code
}
\description{
Gets hexcode from a "color" object.
}
