% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximize_contrast.R
\name{maximize_contrast}
\alias{maximize_contrast}
\title{Finds the highest contrast color available for those specified by the user.
Uses gradient-descent optimization on the "contrast" function.}
\usage{
maximize_contrast(..., maxrep = 10, maxit = 100, h = 1e-08,
  output = "hex")
}
\arguments{
\item{...}{All the colors of interest. Hexcode, c(r,g,b), color 
name as character and "color" class object are all supported.}

\item{maxrep}{Number of times the process of local optimization will be repeated. Defaults to 10}

\item{maxit}{Number of iterations for each gradient descent optimization (finding of local minimum). Defaults to 100}

\item{h}{Step used to calculate the gradient of the contrast function. Defaults to 1e-8}
}
\description{
Finds the highest contrast color available for those specified by the user.
Uses gradient-descent optimization on the "contrast" function.
}
