% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_contrast_colors.R
\name{maximum_contrast_colors}
\alias{maximum_contrast_colors}
\title{Finds as many high-contrast colors as specified by the user.}
\usage{
maximum_contrast_colors(..., n, remove_white = TRUE, seed = 1,
  maxrep = 10, maxit = 100, h = 1e-08, output = "hex")
}
\arguments{
\item{...}{If desired, user-specified colors that must appear on the list.
excode, c(r,g,b), color name as character and "color" class object are all supported.}

\item{n}{Number of colors to be found.}

\item{remove_white}{Prevent white from being added to the list?
Recommended for usage on white backgrounds. Defaults to TRUE.}

\item{seed}{Seed to initialize random number generator. Used
for consistency between executions.}

\item{maxrep}{maxrep parameter for maximize_contrast function call}

\item{maxit}{maxit parameter for maximize_contrast function call}

\item{h}{h parameter for maximize_contrast function call}

\item{hex}{Output format. "hex" and "color" supported.}
}
\description{
Finds as many high-contrast colors as specified by the user.
}
