% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mix}
\alias{mix}
\title{Creates an object of class "color" as a mixture of the colors it receives as input.
TODO: Arranxar o metodo "paint"}
\usage{
mix(..., method = "light", average = method == "paint",
  output = "hex")
}
\arguments{
\item{...}{Colors to be mixed. Hexcode, c(r,g,b), color
name as character and "color" class object are all supported.}

\item{method}{Type of mixture ("light" or "paint") to be made. Defaults to "light"}

\item{average}{If TRUE, the function will return the average color of the inputs. If FALSE,
the function will add the inputs together. When working with "paint" method, average=TRUE is
recommended to represent accurately the mixture of actual paint from idfferent colors.}

\item{output}{Format in which output will be returned. Use "hex" to get a hexcode. Use "color"
for a "color" class object to be returned. Defaults to "hex"}
}
\value{
A hexcode or class "color" object resulting of a mixture of the inputted colors
}
\description{
Creates an object of class "color" as a mixture of the colors it receives as input.
TODO: Arranxar o metodo "paint"
}
\examples{
mix("red", "blue", method="light")
mix("red", "blue", method="paint")
}
