% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_step.R
\name{optimal_step}
\alias{optimal_step}
\title{Calculates best possible step for contrast gradient descent optimization.}
\usage{
optimal_step(x, d, ..., pasos = c(0, 1e-08, 2e-08, 5e-08, 1e-07, 2e-07,
  5e-07, 1e-06, 2e-06, 5e-06, 1e-05, 2e-05, 5e-05, 1e-04, 2e-04, 5e-04,
  0.001, 0.002, 0.005, 0.01, 0.02, 0.05, 0.1, 0.2, 0.5, 1, 2, 5, 10, 20,
  50, 100))
}
\arguments{
\item{x}{Point of origin}

\item{d}{Direction of gradient}

\item{...}{Colors to compare with x (calculate contrast with x)}

\item{pasos}{Candidate steps to choose from}
}
\description{
Calculates best possible step for contrast gradient descent optimization.
}
