% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_color.R
\name{paint_color}
\alias{paint_color}
\title{Creates an object of class "color" from a mixture of paint primary colors (YMCK palette)}
\usage{
paint_color(y, m, c, k = 0, maxVal = 255)
}
\arguments{
\item{y}{Amount of yellow paint}

\item{m}{Amount of magenta paint}

\item{c}{Amount to cyan light}

\item{k}{Amount of black paint. Defaults to 0}

\item{maxVal}{Value assigned to represent maximum amount of paint. Defaults to 255}
}
\description{
Creates an object of class "color" from a mixture of paint primary colors (YMCK palette)
}
