% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb_dist.R
\name{rgb_dist}
\alias{rgb_dist}
\title{Calculates a measure of the contrast between col1 and col2}
\usage{
rgb_dist(col1, col2)
}
\arguments{
\item{col1}{Hexcode, c(r,g,b), color
name as character and "color" class object are all supported.}

\item{col2}{Hexcode, c(r,g,b), color
name as character and "color" class object are all supported.}
}
\description{
Calculates a measure of the contrast between col1 and col2
}
