% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\title{Calculates the minimum contrast between col and the colors specified in ...}
\usage{
contrast(col, ...)
}
\arguments{
\item{col}{The color we want to compare. Hexcode, c(r,g,b), color
name as character and "color" class object are all supported.}

\item{...}{All other colors that col will be compared to. Hexcode, c(r,g,b), color
name as character and "color" class object are all supported.}
}
\value{
A number between 0 and 1, representing a measure for contrast.
}
\description{
Calculates the minimum contrast between col and the colors specified in ...
}
\examples{
contrast("red", "blue")
contrast("red", "blue", "black", "green", "white")
contrast("blue", list("#000000", "brown", c(255, 0, 255), "pink", c(0, 128, 176)))
}
