% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_color2.R
\name{df_color2}
\alias{df_color2}
\title{Appends a column named "color" to a data frame containing hex codes (ready
for usage in a plot) The values are determined by mixing two gradients that depend on two columns
specified by the user.Therefore, the output is able to represent the value of two columns together.
Simultaneous low values for both columns return darker colors.}
\usage{
df_color2(df, column1, column2, output_name = "color",
  max_color1 = "red", max_color2 = "green", min_value1 = min(df[,
  column1]), max_value1 = max(df[, column1]), min_value2 = min(df[,
  column2]), max_value2 = max(df[, column2]))
}
\arguments{
\item{df}{Data frame containing the original data.}

\item{column1}{A string containing the name of the column that will determine color's first gradient.}

\item{column2}{A string containing the name of the column that will determine color's second gradient.}

\item{output_name}{Name for the column in which the output will be stored. Defaults to "color"}

\item{max_color1}{Color to be assigned to the row with the highest value in column1. Defaults to red}

\item{max_color2}{Color to be assigned to the row with the highest value in column2. Defaults to green.}

\item{min_value1}{Allows the user to customize the value for column1 that gets no light assigned. Defaults to the minium of all observations.}

\item{max_value1}{Allows the user to customize the value for column1 that gets maximum light assigned. Defaults to the maximum of all observations.}

\item{min_value2}{Allows the user to customize the value for column1 that gets no light assigned. Defaults to the minium of all observations.}

\item{max_value2}{Allows the user to customize the value for column1 that gets maximum light assigned. Defaults to the maximum of all observations.}
}
\description{
Appends a column named "color" to a data frame containing hex codes (ready
for usage in a plot) The values are determined by mixing two gradients that depend on two columns
specified by the user.Therefore, the output is able to represent the value of two columns together.
Simultaneous low values for both columns return darker colors.
}
