% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hsv.R
\name{get_hsv}
\alias{get_hsv}
\title{Gets HSV specifications from a "color" object}
\usage{
get_hsv(col)
}
\arguments{
\item{col}{An object of class color.}
}
\description{
Gets HSV specifications from a "color" object
}
