% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rgb.R
\name{get_rgb}
\alias{get_rgb}
\title{Gets RGB specifications from a "color" object}
\usage{
get_rgb(col)
}
\arguments{
\item{col}{Color whose RGB is to be returned. Hexcode, c(r,g,b), color
name as character and "color" class object are all supported.}
}
\value{
A vector containing col's RGB specifications.
}
\description{
Gets RGB specifications from a "color" object
}
\examples{
get_rgb("green")
get_rgb(c(255,0,255))
}
