% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsv_color.R
\name{hsv_color}
\alias{hsv_color}
\title{Creates an object of class "color" from HSV palette specifications}
\usage{
hsv_color(h, s, v, vmax = 1, hmax = 1)
}
\arguments{
\item{h}{Hue}

\item{s}{Saturation}

\item{v}{Value}

\item{vmax}{Maximum permitted input for value. Defaults to 1}

\item{hmax}{Perimeter of the colour circle. Defaults to 1}
}
\description{
Creates an object of class "color" from HSV palette specifications
}
