% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsv_weighted_mean.R
\name{hsv_weighted_mean}
\alias{hsv_weighted_mean}
\title{Finds a color between those received as input in HSV palette}
\usage{
hsv_weighted_mean(col1, col2, w1 = 0.5, w2 = 1 - w1, output = "hex")
}
\arguments{
\item{col1}{First color. Hexcode, c(r,g,b), color
name as character and "color" class object are all supported.}

\item{col2}{Second color. Hexcode, c(r,g,b), color
name as character and "color" class object are all supported.}

\item{w1}{Number between 0 and 1. Represents how similar the output will be to col1. Defaults to .5}

\item{w2}{Number between 0 and 1. Represents how similar the output will be to col2. Defaults to 1-w1}

\item{output}{Format in which output will be returned. Use "hex" to get a hexcode. Use "color"
for a "color" class object to be returned. Defaults to "hex"}
}
\value{
A hexcode (or "class" color object) built as commanded.
}
\description{
Finds a color between those received as input in HSV palette
}
\examples{
hsv_weighted_mean("red", "green", output="color")
hsv_weighted_mean("red", "green")
hsv_weighted_mean("blue", "red", w1=5/6)
hsv_weighted_mean(c(0,0,0), color("white"))
hsv_weighted_mean("#000000", "red")
}
