% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantissa.R
\name{mantissa}
\alias{mantissa}
\title{Calculates a number's mantissa (decimal part, or distance to the nearest lower integer)}
\usage{
mantissa(x)
}
\arguments{
\item{x}{Number to calculate mantissa.}
}
\description{
Calculates a number's mantissa (decimal part, or distance to the nearest lower integer)
}
